fix memory corruption in udev_device
udev_device_get_udev directly returns a pointer
to the udev stored udev context in the device
and does not increase the ref count on the context.
Dropping a UdevDevice created from the retrieved context
and device will unref the context which leads to a
memory corruption. This can be prevented by increasing
the ref count of the context prior to creating a UdevDevice.