Fix bitflag tests
The correct idiom is `== <FLAG>`, not `== bitmask`. But that's rather
verbose here. `!= 0` should be fine as long as each of these flags is a
single bit.

This fixes an issue with cosmic-comp where entries in
`~/.local/state/cosmic-comp/inputs.ron` have `scroll_config` values
despite having nothing to do with scrolling.