Generate bindings once for all platforms and architectures
Based on the CI in
https://github.com/Smithay/drm-rs/pull/176.

`libinput_log_set_handler` is disabled, since it requires `va_list`.
That doesn't seem practical to support until `std::ffi::VaList` is
stable.

I don't think BSDs should be different here, but I haven't tested it
yet. CI confirms that the bindings are the same on Linux for tested
architectures.